\name{bipois}
\alias{bipois}
\title{The Bivariate Poisson Distribution}
\description{
Density and random generation for the bivariate Poisson with parameter lambda1, lambda2, lambda3.
}
\usage{
dbipois(x, y, lambda1, lambda2, lambda3 = 0, log = FALSE)
rbipois(n, lambda1, lambda2, lambda3 = 0)
}
\Arguments{
}
\examples{
dbipois(1,0,1.5,1.2,0)
}
